% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_draws.R
\name{fitted_draws.ulam}
\alias{fitted_draws.ulam}
\alias{fitted_draws.quap}
\alias{fitted_draws.map}
\alias{fitted_draws.map2stan}
\title{Add draws from the posterior link-level predictor of a rethinking model to a data frame}
\usage{
\method{fitted_draws}{ulam}(
  model,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  n = NULL,
  seed = NULL,
  scale = c("response", "linear"),
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)

\method{fitted_draws}{quap}(
  model,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  n = NULL,
  seed = NULL,
  scale = c("response", "linear"),
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)

\method{fitted_draws}{map}(
  model,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  n = NULL,
  seed = NULL,
  scale = c("response", "linear"),
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)

\method{fitted_draws}{map2stan}(
  model,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  n = NULL,
  seed = NULL,
  scale = c("response", "linear"),
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)
}
\arguments{
\item{model}{A model fit using \code{rethinking::quap()}, \code{rethinking::ulam()},
\code{rethinking::map()}, or \code{rethinking::map2stan()}.}

\item{newdata}{Data frame to generate predictions from. If omitted, most model types will
generate predictions from the data used to fit the model.}

\item{value}{The name of the output column for \code{fitted_draws}; default \code{".value"}.}

\item{...}{Optional parameters passed on to \code{rethinking::link()}. The most pertinent are:
\itemize{
\item \code{replace}: Optional named list of samples to replace inside posterior samples. See examples in \code{rethinking::link()}.
}}

\item{post}{Optional samples from posterior. When missing, \code{fitted_draws()} extracts these in advance,
bypassing \code{rethinking::link()}'s normal process (\code{rethinking::link()} uses \code{rstan::extract()}, which
unfortunately permutes samples, breaking the ability of the \code{.draw} column to be meaningfully joined
with output from other methods, like \code{spread_draws()}).}

\item{n}{The number of draws per fit to return. When \code{NULL} (the default), \code{rethinking::ulam()} and
\code{rethinking::map2stan()} models return all draws; \code{rethinking::quap()} and \code{rethinking::map()} models
return 5000 draws.}

\item{seed}{A seed to use when subsampling draws (i.e. when \code{n} is not \code{NULL}).}

\item{scale}{Must be \code{"response"} for rethinking models, as the inverse-link function is
always applied.}

\item{dpar}{Should distributional regression
parameters be included in the output? In rethinking models, these correspond to the linear submodels
returned by \code{rethinking::link()}. If \code{TRUE}, distributional regression
parameters are included in the output as additional columns named after each parameter
(alternative names can be provided using a list or named vector, e.g. \code{c(sigma.hat = "sigma")}
would output the \code{"sigma"} parameter from a model as a column named \code{"sigma.hat"}).
If \code{FALSE} (the default), distributional regression parameters are not included; instead,
just the first linear submodel returned by \code{rethinking::link()} is used.}

\item{re_formula, category}{Not used with this model type.}
}
\description{
Adds draws from the posterior link-level predictor of a rethinking model to a data frame.
Provides support for \code{\link[tidybayes:fitted_draws]{tidybayes::fitted_draws()}} / \code{\link[tidybayes:add_fitted_draws]{tidybayes::add_fitted_draws()}} /
\code{\link[tidybayes:add_linpred_draws]{tidybayes::add_linpred_draws()}} for models from the \code{rethinking} package.
}
