% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpred_draws.R
\name{linpred_draws.ulam}
\alias{linpred_draws.ulam}
\alias{linpred_draws.quap}
\alias{linpred_draws.map}
\alias{linpred_draws.map2stan}
\title{Add draws from the posterior link-level predictor of a rethinking model to a data frame}
\usage{
\method{linpred_draws}{ulam}(
  object,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  ndraws = NULL,
  seed = NULL,
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)

\method{linpred_draws}{quap}(
  object,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  ndraws = NULL,
  seed = NULL,
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)

\method{linpred_draws}{map}(
  object,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  ndraws = NULL,
  seed = NULL,
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)

\method{linpred_draws}{map2stan}(
  object,
  newdata,
  value = ".value",
  ...,
  post = NULL,
  ndraws = NULL,
  seed = NULL,
  dpar = FALSE,
  re_formula = NULL,
  category = ".category"
)
}
\arguments{
\item{object}{A model fit using \code{rethinking::quap()}, \code{rethinking::ulam()},
\code{rethinking::map()}, or \code{rethinking::map2stan()}.}

\item{newdata}{Data frame to generate predictions from.}

\item{value}{The name of the output column:
\itemize{
\item for \verb{[add_]epred_draws()}, defaults to \code{".epred"}.
\item for \verb{[add_]predicted_draws()}, defaults to \code{".prediction"}.
\item for \verb{[add_]linpred_draws()}, defaults to \code{".linpred"}.
\item for \verb{[add_]residual_draws()}, defaults to \code{".residual"}
}}

\item{...}{Optional parameters passed on to \code{rethinking::link()}. The most pertinent are:
\itemize{
\item \code{replace}: Optional named list of samples to replace inside posterior samples. See examples in \code{rethinking::link()}.
}}

\item{post}{Optional samples from posterior. When missing, \code{linpred_draws()} extracts these in advance,
bypassing \code{rethinking::link()}'s normal process (\code{rethinking::link()} uses \code{rstan::extract()}, which
unfortunately permutes samples, breaking the ability of the \code{.draw} column to be meaningfully joined
with output from other methods, like \code{spread_draws()}).}

\item{ndraws}{The number of draws per fit to return. When \code{NULL} (the default), \code{rethinking::ulam()} and
\code{rethinking::map2stan()} models return all draws; \code{rethinking::quap()} and \code{rethinking::map()} models
return 5000 draws.}

\item{seed}{A seed to use when subsampling draws (i.e. when \code{ndraws} is not \code{NULL}).}

\item{dpar}{Should distributional regression
parameters be included in the output? In rethinking models, these correspond to the linear submodels
returned by \code{rethinking::link()}. If \code{TRUE}, distributional regression
parameters are included in the output as additional columns named after each parameter
(alternative names can be provided using a list or named vector, e.g. \code{c(sigma.hat = "sigma")}
would output the \code{"sigma"} parameter from a model as a column named \code{"sigma.hat"}).
If \code{FALSE} (the default), distributional regression parameters are not included; instead,
just the first linear submodel returned by \code{rethinking::link()} is used.}

\item{re_formula, category}{Not used with this model type.}
}
\description{
Adds draws from the posterior link-level predictor of a rethinking model to a data frame.
Provides support for \code{\link[tidybayes:add_predicted_draws]{tidybayes::linpred_draws()}} / \code{\link[tidybayes:add_predicted_draws]{tidybayes::add_linpred_draws()}} /
for models from the \code{rethinking} package.
}
