% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_draws.R
\name{predicted_draws.ulam}
\alias{predicted_draws.ulam}
\alias{predicted_draws.quap}
\alias{predicted_draws.map}
\alias{predicted_draws.map2stan}
\title{Add draws from the posterior predictive of a rethinking model to a data frame}
\usage{
\method{predicted_draws}{ulam}(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

\method{predicted_draws}{quap}(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

\method{predicted_draws}{map}(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

\method{predicted_draws}{map2stan}(
  model,
  newdata,
  prediction = ".prediction",
  ...,
  n = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)
}
\arguments{
\item{model}{A model fit using \code{rethinking::quap()}, \code{rethinking::ulam()},
\code{rethinking::map()}, or \code{rethinking::map2stan()}.}

\item{newdata}{Data frame to generate predictions from. If omitted, most model types will
generate predictions from the data used to fit the model.}

\item{prediction}{The name of the output column for \code{predicted_draws}; default \code{".prediction"}.}

\item{...}{Optional parameters passed on to \code{rethinking::sim()}. The most pertinent are:
\itemize{
\item \code{post}: Optional samples from posterior. If missing, simulates samples using \code{n}.
}}

\item{n}{The number of draws per prediction to return. When \code{NULL} (the default), \code{rethinking::ulam()} and
\code{rethinking::map2stan()} models return all draws; \code{rethinking::quap()} and \code{rethinking::map()} models
return 5000 draws.}

\item{seed}{A seed to use when subsampling draws (i.e. when \code{n} is not \code{NULL}).}

\item{re_formula, category}{Not used with this model type.}
}
\description{
Adds draws from the posterior predictive distribution of a rethinking model to a data frame.
Provides support for \code{\link[tidybayes:predicted_draws]{tidybayes::predicted_draws()}} / \code{\link[tidybayes:add_predicted_draws]{tidybayes::add_predicted_draws()}} for
models from the \code{rethinking} package.
}
