% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_draws.R
\name{tidy_draws.map}
\alias{tidy_draws.map}
\alias{tidy_draws.quap}
\title{Get a sample of posterior draws from quap and map models as a tibble}
\usage{
\method{tidy_draws}{map}(model, n = 10000, ...)

\method{tidy_draws}{quap}(model, n = 10000, ...)
}
\description{
Implementation of \code{\link[tidybayes:tidy_draws]{tidybayes::tidy_draws()}} for \code{\link[rethinking:map]{rethinking::map()}} and \code{\link[rethinking:quap]{rethinking::quap()}}
models. Extract draws from a Bayesian fit into a wide-format data frame with a
\code{.chain}, \code{.iteration}, and \code{.draw} column, as well as all variables as columns.
While this function can be useful for quick glances at models (especially
combined with \code{\link[=gather_variables]{gather_variables()}} and \code{\link[=median_qi]{median_qi()}}), it is
generally speaking not as useful as \code{\link[=spread_draws]{spread_draws()}} or
\code{\link[=gather_draws]{gather_draws()}} for most applications, and is mainly used internally.
}
\details{
The main additional functionality compared to \code{\link[tidybayes:tidy_draws]{tidybayes::tidy_draws()}} when used on
other models is that since draws must be generated on-the-fly,
an argument (\code{n}) is provided to indicate how many draws to take. The \code{.chain} and
\code{.iteration} columns are also always \code{NA}, since they have no meaning for these model
types (use the \code{.draw} column if you need to index draws). Otherwise, the result of
this function follows the same format as \code{\link[tidybayes:tidy_draws]{tidybayes::tidy_draws()}}; see that
documentation for more information.
}
\examples{

library(rethinking)
library(tidybayes)
library(magrittr)

m = quap(alist(
    mpg ~ dlnorm(mu, sigma),
    mu <- a + b*wt,
    c(a,b) ~ dnorm(0, 10),
    sigma ~ dexp(1)
  ),
  data = mtcars,
  start = list(a = 4, b = -1, sigma = 1)
)

m \%>\%
  tidy_draws() \%>\%
  gather_variables() \%>\%
  median_qi()

}
