% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/fitted_draws.R,
%   R/predicted_draws.R
\name{tidybayes.rethinking-deprecated}
\alias{tidybayes.rethinking-deprecated}
\alias{tidy_link}
\alias{tidy_sim}
\title{Deprecated functions, arguments, and column names in tidybayes}
\usage{
tidy_link(data, fit, ...)

tidy_sim(data, fit, ...)
}
\description{
Deprecated functions, arguments, and column names and their alternatives are listed below.
}
\section{Deprecated Functions}{


When \code{tidybayes.rethinking} was originally created, \code{tidybayes} did not have good
analogs for \code{tidy_link()} or \code{tidy_sim()}. However, \code{tidybayes} now has a mature API
analogous to these functions, so they have been deprecated:
\itemize{
\item \code{tidy_link()} has been replaced with \code{add_fitted_draws()} / \code{fitted_draws()}. See
\code{fitted_draws.ulam()} for information on rethinking-specific arguments, or
\code{tidybayes::fitted_draws()} for information on the general function.
\item \code{tidy_sim()} has been replaced with \code{add_predicted_draws()} / \code{predicted_draws()}. See
\code{predicted_draws.ulam()} for information on rethinking-specific arguments, or
\code{tidybayes::predicted_draws()} for information on the general function.
}
}

\author{
Matthew Kay
}
