% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_draws.R
\name{predicted_draws.ulam}
\alias{predicted_draws.ulam}
\alias{predicted_draws.quap}
\alias{predicted_draws.map}
\alias{predicted_draws.map2stan}
\title{Add draws from the posterior predictive of a rethinking model to a data frame}
\usage{
\method{predicted_draws}{ulam}(
  object,
  newdata,
  value = ".prediction",
  ...,
  ndraws = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

\method{predicted_draws}{quap}(
  object,
  newdata,
  value = ".prediction",
  ...,
  ndraws = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

\method{predicted_draws}{map}(
  object,
  newdata,
  value = ".prediction",
  ...,
  ndraws = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)

\method{predicted_draws}{map2stan}(
  object,
  newdata,
  value = ".prediction",
  ...,
  ndraws = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category"
)
}
\arguments{
\item{object}{A model fit using \code{rethinking::quap()}, \code{rethinking::ulam()},
\code{rethinking::map()}, or \code{rethinking::map2stan()}.}

\item{newdata}{Data frame to generate predictions from.}

\item{value}{The name of the output column:
\itemize{
\item for \verb{[add_]epred_draws()}, defaults to \code{".epred"}.
\item for \verb{[add_]predicted_draws()}, defaults to \code{".prediction"}.
\item for \verb{[add_]linpred_draws()}, defaults to \code{".linpred"}.
\item for \verb{[add_]residual_draws()}, defaults to \code{".residual"}
}}

\item{...}{Optional parameters passed on to \code{rethinking::sim()}. The most pertinent are:
\itemize{
\item \code{post}: Optional samples from posterior. If missing, simulates samples using \code{ndraws}.
}}

\item{ndraws}{The number of draws per prediction to return. When \code{NULL} (the default), \code{rethinking::ulam()} and
\code{rethinking::map2stan()} models return all draws; \code{rethinking::quap()} and \code{rethinking::map()} models
return 5000 draws.}

\item{seed}{A seed to use when subsampling draws (i.e. when \code{ndraws} is not \code{NULL}).}

\item{re_formula, category}{Not used with this model type.}
}
\description{
Adds draws from the posterior predictive distribution of a rethinking model to a data frame.
Provides support for \code{\link[tidybayes:add_predicted_draws]{tidybayes::predicted_draws()}} / \code{\link[tidybayes:add_predicted_draws]{tidybayes::add_predicted_draws()}} for
models from the \code{rethinking} package.
}
